Update contenttokens.php
I love this extension, and I wanted to use it with Drupal 8 (I am running CiviCRM with Drupal 8.5). I couldn't find an existing update for it, so I have amended this file to accommodate the Drupal 8 use case. I tried to avoid breaking anything, but I have not tested this is in anything other than Drupal 8. It appears to work in D8, at least up to the point of tokens showing up in the menu, and placed in the body or subject of an email. So far, I haven't been able to do a test send. I have commented every change; feel free to remove them, or I can at your request.